refactor: 퇴원, 면회 관련 폴더구조 파일명 변경
