+++ components/hospital/icu/main/out-and-visit/icu-out-chart/icu-out-chart.tsx
@@ -1,7 +1,7 @@
 import NoResultSquirrel from '@/components/common/no-result-squirrel'
 import DataTable from '@/components/ui/data-table'
-import { getIcuOutDuePatients } from '@/lib/services/icu/movement/out-due/get-out-due-patients'
 import { outDueColumns } from './out-due-columns'
+import { getIcuOutDuePatients } from '@/lib/services/icu/out-and-visit/icu-out-chart'
 
 export default async function IcuOutChart({
   hosId,
