--- lib/services/icu/movement/out-due/get-out-due-patients.ts
@@ -1,26 +1,0 @@
-'use server'
-
-import { createClient } from '@/lib/supabase/server'
-import type { OutDuePatientsData } from '@/types/icu/movement'
-import { redirect } from 'next/navigation'
-
-export const getIcuOutDuePatients = async (
-  hosId: string,
-  targetDate: string,
-) => {
-  const supabase = createClient()
-
-  const { data, error } = await supabase
-    .rpc('get_icu_out_due_patients', {
-      hos_id_input: hosId,
-      target_date_input: targetDate,
-    })
-    .returns<OutDuePatientsData[]>()
-
-  if (error) {
-    console.error(error)
-    redirect(`/error?message=${error?.message}`)
-  }
-
-  return data || []
-}
