+++ lib/services/icu/out-and-visit/visit-chart.ts
@@ -25,3 +25,19 @@ export const insertVisitPatient = async (
     redirect(`/error?message=${error?.message}`)
   }
 }
+
+export const updateIsVisitDone = async (visitId: string, isDone: boolean) => {
+  const supabase = createClient()
+
+  const { error } = await supabase
+    .from('icu_visit')
+    .update({
+      is_done: isDone,
+    })
+    .match({ icu_visit_id: visitId })
+
+  if (error) {
+    console.error(error)
+    redirect(`/error?message=${error?.message}`)
+  }
+}
